Bij nader inzien, dit hoeft niet. Er wordt namelijk gekeken of er een doos(2) in de matrix te vinden is. Dat is alleen mogelijk als een doos NIET op een doellocatie staat. Dit staat los van het feit of er een doelvak over is of niet.
